move bokeh.util.testing to bokeh.testing.util
