+++ bokeh/application/tests/test___init__.py
@@ -22,7 +22,7 @@ import pytest ; pytest
 # External imports
 
 # Bokeh imports
-from bokeh.util.testing import verify_all
+from bokeh.testing.util import verify_all
 
 # Module under test
 import bokeh.application as ba
