+++ bokeh/command/subcommands/tests/test_svg.py
@@ -9,7 +9,7 @@ is_python2 = sys.version_info[0] == 2
 
 import bokeh.command.subcommands.svg as scsvg
 from bokeh.command.bootstrap import main
-from bokeh.util.testing import TmpDir, WorkingDir, with_directory_contents
+from bokeh.testing.util import TmpDir, WorkingDir, with_directory_contents
 
 from . import basic_svg_scatter_script, multi_svg_scatter_script
