+++ bokeh/sampledata/tests/test_unemployment1948.py
@@ -22,8 +22,8 @@ import pytest ; pytest
 # External imports
 
 # Bokeh imports
-from bokeh.util.testing import verify_all
 from bokeh.testing.fixtures import pd; pd
+from bokeh.testing.util import verify_all
 
 # Module under test
 #import bokeh.sampledata.unemployment1948 as bsu
