+++ bokeh/sampledata/tests/test_us_counties.py
@@ -22,7 +22,7 @@ import pytest ; pytest
 # External imports
 
 # Bokeh imports
-from bokeh.util.testing import verify_all
+from bokeh.testing.util import verify_all
 
 # Module under test
 #import bokeh.sampledata.us_counties as bsu
