+++ bokeh/tests/test___main__.py
@@ -23,7 +23,7 @@ from mock import patch
 # External imports
 
 # Bokeh imports
-from bokeh.util.testing import verify_all
+from bokeh.testing.util import verify_all
 
 # Module under test
 import bokeh.__main__ as bm
