+++ bokeh/tests/test_tile_providers.py
@@ -23,7 +23,7 @@ import pytest ; pytest
 
 # Bokeh imports
 from bokeh.models.tiles import WMTSTileSource
-from bokeh.util.testing import verify_all
+from bokeh.testing.util import verify_all
 
 # Module under test
 import bokeh.tile_providers as bt
