+++ bokeh/util/__init__.py
@@ -102,16 +102,6 @@ functionality.
 .. automodule:: bokeh.util.string
   :members:
 
-.. _bokeh.util.testing:
-
-``bokeh.util.testing``
-----------------------
-
-.. automodule:: bokeh.util.testing
-  :members:
-
-.. _bokeh.util.tornado:
-
 ``bokeh.util.tornado``
 ----------------------
